NIDAQ: Split out AnalogOutputPort in an individual file.
