+++ DeviceAdapters/NIDAQ/NIDAQ.vcxproj.filters
@@ -18,6 +18,9 @@
     <ClCompile Include="NIDAQ.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
+    <ClCompile Include="NIAnalogOutputPort.cpp">
+      <Filter>Source Files</Filter>
+    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="NIDAQ.h">
